Created a new component for password strength checking.
